test(`cqn4sql`): consolidate calc-element specs
Rename other-places.spec.js → outside-select-list.spec.js and fold the
two exists tests into it as a nested describe — they're a special case
of calc elements unfolding outside the select list, not a standalone
topic worth its own file.
